# .ai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wallet API",
    "description": "Sistema de Gestão Financeira",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/transactions": {
      "get": {
        "tags": [
          "Transações e Fluxo"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_api_v1_transactions_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por PESSOAL ou EMPRESA",
              "title": "Profile"
            },
            "description": "Filtrar por PESSOAL ou EMPRESA"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "RECEITA ou DESPESA",
              "title": "Type"
            },
            "description": "RECEITA ou DESPESA"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "PENDENTE, CONCLUIDO, CANCELADO",
              "title": "Status"
            },
            "description": "PENDENTE, CONCLUIDO, CANCELADO"
          },
          {
            "name": "start_due_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Vencimento inicial (YYYY-MM-DD)",
              "title": "Start Due Date"
            },
            "description": "Vencimento inicial (YYYY-MM-DD)"
          },
          {
            "name": "end_due_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Vencimento final (YYYY-MM-DD)",
              "title": "End Due Date"
            },
            "description": "Vencimento final (YYYY-MM-DD)"
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por Categoria",
              "title": "Category Id"
            },
            "description": "Filtrar por Categoria"
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por Conta Bancária / Carteira",
              "title": "Account Id"
            },
            "description": "Filtrar por Conta Bancária / Carteira"
          },
          {
            "name": "contact_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por Contato / Favorecido",
              "title": "Contact Id"
            },
            "description": "Filtrar por Contato / Favorecido"
          },
          {
            "name": "debt_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por Dívida vinculada",
              "title": "Debt Id"
            },
            "description": "Filtrar por Dívida vinculada"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Busca textual na descrição ou observações",
              "title": "Search"
            },
            "description": "Busca textual na descrição ou observações"
          },
          {
            "name": "is_overdue",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar apenas contas em atraso (vencidas e não pagas)",
              "title": "Is Overdue"
            },
            "description": "Filtrar apenas contas em atraso (vencidas e não pagas)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response List Transactions Api V1 Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transações e Fluxo"
        ],
        "summary": "Create Transaction",
        "operationId": "create_transaction_api_v1_transactions_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{trans_id}": {
      "get": {
        "tags": [
          "Transações e Fluxo"
        ],
        "summary": "Get Transaction",
        "operationId": "get_transaction_api_v1_transactions__trans_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trans_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trans Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Transações e Fluxo"
        ],
        "summary": "Update Transaction",
        "operationId": "update_transaction_api_v1_transactions__trans_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trans_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trans Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transações e Fluxo"
        ],
        "summary": "Delete Transaction",
        "operationId": "delete_transaction_api_v1_transactions__trans_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trans_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trans Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{trans_id}/complete": {
      "patch": {
        "tags": [
          "Transações e Fluxo"
        ],
        "summary": "Complete Transaction",
        "description": "Marca uma transação pendente como CONCLUIDA e abate dívidas vinculadas.",
        "operationId": "complete_transaction_api_v1_transactions__trans_id__complete_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trans_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trans Id"
            }
          },
          {
            "name": "payment_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data de pagamento (YYYY-MM-DD), padrão hoje",
              "title": "Payment Date"
            },
            "description": "Data de pagamento (YYYY-MM-DD), padrão hoje"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransactionCreate": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "RECEITA",
              "DESPESA"
            ],
            "title": "Type",
            "description": "RECEITA ou DESPESA"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id",
            "description": "ID da categoria"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "contact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "debt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt Id"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Description"
          },
          "amount_cents": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Cents",
            "description": "Valor em centavos"
          },
          "due_date": {
            "type": "string",
            "title": "Due Date",
            "description": "Data prevista / vencimento (YYYY-MM-DD)"
          },
          "payment_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Date",
            "description": "Data de quitação real (YYYY-MM-DD)"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "CONCLUIDO",
              "CANCELADO"
            ],
            "title": "Status",
            "default": "PENDENTE"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "profile": {
            "type": "string",
            "enum": [
              "PESSOAL",
              "EMPRESA"
            ],
            "title": "Profile",
            "description": "Perfil da transação"
          },
          "attachment_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachment Ids",
            "description": "IDs dos anexos previamente carregados para vincular"
          }
        },
        "type": "object",
        "required": [
          "type",
          "category_id",
          "description",
          "amount_cents",
          "due_date",
          "profile"
        ],
        "title": "TransactionCreate"
      },
      "TransactionResponse": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "RECEITA",
              "DESPESA"
            ],
            "title": "Type",
            "description": "RECEITA ou DESPESA"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "category_id": {
            "type": "string",
            "title": "Category Id",
            "description": "ID da categoria"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "contact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "debt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt Id"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Description"
          },
          "amount_cents": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Cents",
            "description": "Valor em centavos"
          },
          "due_date": {
            "type": "string",
            "title": "Due Date",
            "description": "Data prevista / vencimento (YYYY-MM-DD)"
          },
          "payment_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Date",
            "description": "Data de quitação real (YYYY-MM-DD)"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "CONCLUIDO",
              "CANCELADO"
            ],
            "title": "Status",
            "default": "PENDENTE"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "profile": {
            "type": "string",
            "enum": [
              "PESSOAL",
              "EMPRESA"
            ],
            "title": "Profile"
          },
          "schedule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "installment_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Number"
          },
          "total_installments": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Installments"
          },
          "sync_status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "SINCRONIZADO"
            ],
            "title": "Sync Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentResponse"
            },
            "type": "array",
            "title": "Attachments"
          },
          "attachments_count": {
            "type": "integer",
            "title": "Attachments Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "type",
          "category_id",
          "description",
          "amount_cents",
          "due_date",
          "id",
          "profile",
          "sync_status",
          "created_at",
          "updated_at"
        ],
        "title": "TransactionResponse"
      },
      "TransactionUpdate": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "contact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "debt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debt Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "amount_cents": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Cents"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "payment_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Date"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "PENDENTE",
                  "CONCLUIDO",
                  "CANCELADO"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "TransactionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}